Split out slurm related code from data prep
